ref(backend): Migrate from Express to Hono + Bun
Move backend from root-level Express/Socket.IO to dedicated codesync-server/ directory using Hono + Bun for improved performance and simpler deployment.

- Create codesync-server/ with Hono HTTP server, Socket.IO integration, and TypeScript config
- Remove root-level server.ts, action.js, tsconfig.prod.json and Express/cors/nodemon deps
- Update Editor.tsx import path from ../../action to @/utils/constants
- Clean root package.json scripts and dependencies
