+++ codesync-server/.env.example
@@ -1,0 +1,7 @@
+PORT=3000
+
+# CORS origins (comma-separated). Use * to allow all (not recommended for production).
+CORS_ORIGIN=http://localhost:5173
+
+# Example for production:
+# CORS_ORIGIN=https://your-frontend.vercel.app,https://your-custom-domain.com
