+++ codesync-server/src/socket.ts
@@ -1,0 +1,408 @@
+import type { Server as HttpServer } from "node:http";
+import type { Socket } from "socket.io";
+import { Server as SocketServer } from "socket.io";
+import { ACTIONS } from "./actions.js";
+
+type UserPermissions = {
+  canEdit: boolean;
+  canCreateTab: boolean;
+  canDeleteTab: boolean;
+  canRenameTab: boolean;
+};
+
+type TabData = { name: string; code: string };
+
+const ROOM_CLEANUP_DELAY_MS = 500;
+const DEFAULT_TAB_ID = "tab-main";
+const DEFAULT_TAB_NAME = "main.js";
+
+const DEFAULT_PERMISSIONS: UserPermissions = {
+  canEdit: false,
+  canCreateTab: false,
+  canDeleteTab: false,
+  canRenameTab: false,
+};
+
+const OWNER_PERMISSIONS: UserPermissions = {
+  canEdit: true,
+  canCreateTab: true,
+  canDeleteTab: true,
+  canRenameTab: true,
+};
+
+export function setupSocket(
+  httpServer: HttpServer,
+  isAllowedOrigin: (origin: string | undefined) => boolean
+): SocketServer {
+  const io = new SocketServer(httpServer, {
+    cors: {
+      origin: (origin, callback) => {
+        if (isAllowedOrigin(origin)) {
+          callback(null, true);
+        } else {
+          callback(new Error("CORS policy: This origin is not allowed"));
+        }
+      },
+      methods: ["GET", "POST"],
+      credentials: true,
+    },
+  });
+
+  const userSocketMap = new Map<string, string>();
+  const roomCreatorMap = new Map<string, string>();
+  const roomTabsMap = new Map<string, Map<string, TabData>>();
+  const userActiveTabMap = new Map<string, string>();
+  const roomCurrentEditorMap = new Map<string, string>();
+  const roomPermissionsMap = new Map<string, Map<string, UserPermissions>>();
+
+  const getOrCreateRoomTabs = (roomId: string): Map<string, TabData> => {
+    let tabs = roomTabsMap.get(roomId);
+    if (!tabs) {
+      tabs = new Map<string, TabData>();
+      tabs.set(DEFAULT_TAB_ID, { name: DEFAULT_TAB_NAME, code: "" });
+      roomTabsMap.set(roomId, tabs);
+    }
+    return tabs;
+  };
+
+  const getOrCreateRoomPermissions = (
+    roomId: string
+  ): Map<string, UserPermissions> => {
+    let perms = roomPermissionsMap.get(roomId);
+    if (!perms) {
+      perms = new Map<string, UserPermissions>();
+      roomPermissionsMap.set(roomId, perms);
+    }
+    return perms;
+  };
+
+  const getRoomCurrentEditor = (roomId: string): string =>
+    roomCurrentEditorMap.get(roomId) ?? "";
+
+  const setRoomCurrentEditor = (roomId: string, editor: string) => {
+    roomCurrentEditorMap.set(roomId, editor);
+  };
+
+  const serializeTabs = (
+    tabs: Map<string, TabData>
+  ): { id: string; name: string; code: string }[] =>
+    [...tabs.entries()].map(([id, { name, code }]) => ({ id, name, code }));
+
+  const serializePermissions = (
+    perms: Map<string, UserPermissions>
+  ): Record<string, UserPermissions> => Object.fromEntries(perms);
+
+  const getUserActiveTabs = (
+    roomId: string
+  ): { username: string; activeTabId: string }[] => {
+    const room = io.sockets.adapter.rooms.get(roomId);
+    if (!room) return [];
+    return [...room]
+      .map((sid) => ({
+        username: userSocketMap.get(sid) ?? "",
+        activeTabId: userActiveTabMap.get(sid) ?? DEFAULT_TAB_ID,
+      }))
+      .filter((u) => u.username);
+  };
+
+  const getAllConnectedClients = (roomId: string) =>
+    [...(io.sockets.adapter.rooms.get(roomId) ?? [])].map((socketId) => ({
+      socketId,
+      username: userSocketMap.get(socketId),
+    }));
+
+  io.on("connection", (socket: Socket) => {
+    socket.on(
+      ACTIONS.JOIN,
+      ({ roomId, userName }: { roomId: string; userName: string }) => {
+        userSocketMap.set(socket.id, userName);
+        socket.join(roomId);
+
+        let roomCreator: string;
+        if (roomCreatorMap.has(roomId)) {
+          roomCreator = roomCreatorMap.get(roomId) as string;
+        } else {
+          roomCreatorMap.set(roomId, userName);
+          roomCreator = userName;
+        }
+
+        const clients = getAllConnectedClients(roomId);
+        if (
+          clients.length > 1 &&
+          clients.filter((c) => c.username === userName).length > 1
+        ) {
+          userSocketMap.delete(socket.id);
+          socket.emit(ACTIONS.DUPLICATE_USER, { username: userName });
+          socket.leave(roomId);
+          socket.disconnect();
+          return;
+        }
+
+        const tabs = getOrCreateRoomTabs(roomId);
+        const perms = getOrCreateRoomPermissions(roomId);
+        if (!roomCurrentEditorMap.has(roomId)) {
+          setRoomCurrentEditor(roomId, "");
+        }
+
+        if (roomCreator === userName) {
+          perms.set(userName, { ...OWNER_PERMISSIONS });
+        } else if (!perms.has(userName)) {
+          perms.set(userName, { ...DEFAULT_PERMISSIONS });
+        }
+
+        userActiveTabMap.set(socket.id, DEFAULT_TAB_ID);
+
+        for (const { socketId } of clients) {
+          io.to(socketId).emit(ACTIONS.JOINED, {
+            clients,
+            username: userName,
+            socketId: socket.id,
+            roomcreator: roomCreator,
+          });
+        }
+
+        socket.emit(ACTIONS.TAB_SYNC, {
+          tabs: serializeTabs(tabs),
+          activeTabId: DEFAULT_TAB_ID,
+          userActiveTabs: getUserActiveTabs(roomId),
+          permissions: serializePermissions(perms),
+        });
+      }
+    );
+
+    socket.on(
+      ACTIONS.CODE_CHANGE,
+      ({
+        roomId,
+        tabId,
+        code,
+      }: {
+        roomId: string;
+        tabId: string;
+        code: string;
+      }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        const perms = getOrCreateRoomPermissions(roomId);
+        const userPerms = perms.get(userName) ?? DEFAULT_PERMISSIONS;
+        const currentEditor = getRoomCurrentEditor(roomId);
+        const canEdit = roomCreator === userName || userPerms.canEdit;
+
+        if (!canEdit) return;
+        if (roomCreator !== userName && currentEditor !== userName) return;
+
+        const tabs = roomTabsMap.get(roomId);
+        if (tabs) {
+          const tab = tabs.get(tabId);
+          if (tab) tab.code = code;
+        }
+
+        socket.in(roomId).emit(ACTIONS.CODE_CHANGE, {
+          tabId,
+          code,
+          currenteditor: getRoomCurrentEditor(roomId),
+        });
+      }
+    );
+
+    socket.on(
+      ACTIONS.SYNC_CODE,
+      ({
+        socketId,
+        code,
+        currenteditor,
+        tabId,
+      }: {
+        socketId: string;
+        code: string;
+        currenteditor: string;
+        tabId: string;
+      }) => {
+        io.to(socketId).emit(ACTIONS.CODE_CHANGE, { tabId, code, currenteditor });
+      }
+    );
+
+    socket.on(
+      ACTIONS.TAB_CODE_REQUEST,
+      ({ roomId, tabId }: { roomId: string; tabId: string }) => {
+        if (!socket.rooms.has(roomId)) return;
+        const tabs = roomTabsMap.get(roomId);
+        if (!tabs) return;
+        const tab = tabs.get(tabId);
+        if (!tab) return;
+        socket.emit(ACTIONS.TAB_CODE, { tabId, code: tab.code });
+      }
+    );
+
+    socket.on(
+      ACTIONS.SET_CURRENT_EDITOR,
+      ({
+        roomId,
+        currenteditor,
+      }: {
+        roomId: string;
+        currenteditor: string;
+      }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        const currentEditor = getRoomCurrentEditor(roomId);
+        const isOwner = roomCreator === userName;
+        const canRelease = currenteditor === "" && currentEditor === userName;
+
+        if (!(isOwner || canRelease)) return;
+
+        setRoomCurrentEditor(roomId, currenteditor);
+        socket.in(roomId).emit(ACTIONS.SET_CURRENT_EDITOR, { currenteditor });
+      }
+    );
+
+    socket.on(
+      ACTIONS.TAB_CREATE,
+      ({
+        roomId,
+        tabId,
+        name,
+      }: {
+        roomId: string;
+        tabId: string;
+        name: string;
+      }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        const perms = getOrCreateRoomPermissions(roomId);
+        const userPerms = perms.get(userName) ?? DEFAULT_PERMISSIONS;
+        if (roomCreator !== userName && !userPerms.canCreateTab) return;
+
+        const tabs = roomTabsMap.get(roomId);
+        if (tabs) tabs.set(tabId, { name, code: "" });
+
+        io.in(roomId).emit(ACTIONS.TAB_CREATE, { tabId, name });
+      }
+    );
+
+    socket.on(
+      ACTIONS.TAB_CLOSE,
+      ({ roomId, tabId }: { roomId: string; tabId: string }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        const perms = getOrCreateRoomPermissions(roomId);
+        const userPerms = perms.get(userName) ?? DEFAULT_PERMISSIONS;
+        if (roomCreator !== userName && !userPerms.canDeleteTab) return;
+
+        const tabs = roomTabsMap.get(roomId);
+        if (tabs && tabs.size > 1) {
+          tabs.delete(tabId);
+          io.in(roomId).emit(ACTIONS.TAB_CLOSE, { tabId });
+        }
+      }
+    );
+
+    socket.on(
+      ACTIONS.TAB_RENAME,
+      ({
+        roomId,
+        tabId,
+        name,
+      }: {
+        roomId: string;
+        tabId: string;
+        name: string;
+      }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        const perms = getOrCreateRoomPermissions(roomId);
+        const userPerms = perms.get(userName) ?? DEFAULT_PERMISSIONS;
+        if (roomCreator !== userName && !userPerms.canRenameTab) return;
+
+        const tabs = roomTabsMap.get(roomId);
+        if (tabs) {
+          const tab = tabs.get(tabId);
+          if (tab) tab.name = name;
+        }
+
+        io.in(roomId).emit(ACTIONS.TAB_RENAME, { tabId, name });
+      }
+    );
+
+    socket.on(
+      ACTIONS.TAB_SWITCH,
+      ({ roomId, tabId }: { roomId: string; tabId: string }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        userActiveTabMap.set(socket.id, tabId);
+        socket.in(roomId).emit(ACTIONS.TAB_SWITCH, { username: userName, tabId });
+      }
+    );
+
+    socket.on(
+      ACTIONS.PERMISSIONS_UPDATE,
+      ({
+        roomId,
+        username,
+        permissions,
+      }: {
+        roomId: string;
+        username: string;
+        permissions: UserPermissions;
+      }) => {
+        const userName = userSocketMap.get(socket.id);
+        if (!userName) return;
+
+        const roomCreator = roomCreatorMap.get(roomId);
+        if (roomCreator !== userName) return;
+
+        const perms = getOrCreateRoomPermissions(roomId);
+        perms.set(username, permissions);
+        io.in(roomId).emit(ACTIONS.PERMISSIONS_UPDATE, { username, permissions });
+      }
+    );
+
+    socket.on("disconnecting", () => {
+      const rooms = [...socket.rooms];
+
+      for (const room of rooms) {
+        socket.in(room).emit(ACTIONS.DISCONNECTED, {
+          socketId: socket.id,
+          username: userSocketMap.get(socket.id),
+        });
+
+        setTimeout(() => {
+          if (!io.sockets.adapter.rooms.get(room)) {
+            roomCreatorMap.delete(room);
+            roomTabsMap.delete(room);
+            roomPermissionsMap.delete(room);
+            roomCurrentEditorMap.delete(room);
+          }
+        }, ROOM_CLEANUP_DELAY_MS);
+      }
+
+      const userName = userSocketMap.get(socket.id);
+      if (userName) {
+        for (const room of rooms) {
+          if (getRoomCurrentEditor(room) === userName) {
+            setRoomCurrentEditor(room, "");
+            socket.in(room).emit(ACTIONS.SET_CURRENT_EDITOR, {
+              currenteditor: "",
+            });
+          }
+        }
+      }
+
+      userActiveTabMap.delete(socket.id);
+      userSocketMap.delete(socket.id);
+    });
+  });
+
+  return io;
+}
