+++ package.json
@@ -4,15 +4,10 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev:front": "vite",
-    "dev": "npm run build:frontend && npm run server:prod",
-    "build": "npm run build:frontend && npm run build:backend",
-    "build:frontend": "vite-react-ssg build",
-    "build:backend": "mkdir -p dist && tsc server.ts --outDir dist --module esnext --target es2020 --moduleResolution bundler --skipLibCheck --types node && cp action.js dist/action.js && cp package.json dist/package.json",
+    "dev": "vite",
+    "build": "vite-react-ssg build",
     "build:ssg": "vite-react-ssg build",
     "preview": "vite preview",
-    "server": "nodemon server.ts",
-    "server:prod": "node dist/server.js",
     "check": "ultracite check",
     "fix": "ultracite fix",
     "type-check": "tsc --noEmit",
@@ -32,8 +27,6 @@
     "axios": "^1.7.7",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cors": "^2.8.5",
-    "express": "^4.21.1",
     "lucide-react": "^0.544.0",
     "randomcolor": "^0.6.2",
     "react": "^18.3.1",
@@ -42,7 +35,6 @@
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.28.0",
-    "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.13",
@@ -51,14 +43,11 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.2.4",
-    "@types/cors": "^2.8.19",
-    "@types/express": "^5.0.3",
     "@types/node": "^24.5.2",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
     "globals": "^15.11.0",
-    "nodemon": "^3.1.7",
     "tw-animate-css": "^1.3.8",
     "typescript": "^5.9.2",
     "ultracite": "^5.4.4",
